Issue 31 doc - important fix at hash operation (#101)
* Fix on query dump issue

* Fix for graphviz

* poprawki

* svg and transparent

* fix - additional parameter

* style fixes - cformat

* Important fix on DESC grammar parser

* fix style

* poprawki

* fix test

* Check  method: blobless

* fix

* plotblock

* update doc

* plotblock cleanup

* plot-dsp fix for re-run

* fixes

* fix command and colon

* work on plot

* cleanup scripts

* fix simple-max

* plot-simple to xplot

* cformat

* doc-examples

* fix legend of gnuplot

* fix legend on gnuplot

* doc examples, style fixes

* deb

* fix ci

* test fix ci

* smoke ci fix

* paths

* fix

* xretractor on path

* path

* bump ver

* fixing build env local

* coption - pre

* fix

* coption reduction

* bump version

* fix and example

* Work in Progress - fix for ugly bug in hash operation (delta)

* marble-diagrams-impl

* fix and swritl data generator

* style

* examples and func

* yaml as qxry -t

* uttest fix - now it is yaml

* yaml - fix

* agse  work

* ignore desc

* Docker bump boost and spdlog

* Work on file rotation on close

* cleanup temp and start dev of persistent conter

* cformat

* percounter work

* pragma once

* move cnt

* cleanup procedure in xretractor

* missing cleanup fix

* cleanup <-> rotate

* percountrer impl

* test area ignore

* coption -> percounter in compiler directive

* support for xtrdb

* percounter -> rotation and remove switch

* rotation-test

* comments in descriptor

* percounter -> rotation

* fixing test cleanup

* loop in rql intro

* disposable

* oneshot and disposable

* coreinstace_ in dataModel

* functional disposable

* Disposable and OneShot sync

* move program options var to executor

* get rid off boost::thread - using std, start work on xqrywait

* xqrywait option into

* dead code elimination

* wait for query option intro

* test wait for completion

* service.lock clean

* test optimization time

* separation of zerostep

* Cleanup temp in dsp

* unused var

* small steps - refactor

* cleanup - small steps

* small step refactor

* fetchPayload - cleanup

* small step

* remove of fetchPayload

* simplify processRows

* state lock - removed

* broadcast

* zerostep clean

* work on dead paths

* techtebt payed.

* combined -k and -x on xqry

* fix

* Important fix on stream algebra

+++ src/qry/qryLauncher.cpp
@@ -95,7 +95,7 @@ int main(int argc, char *argv[]) {
     }
     if (vm.count("hello"))
       return obj.hello();
-    else if (vm.count("kill")) {
+    else if (vm.count("kill") && timeLimit == 0) {
       ptree pt = obj.netClient("kill", "");
       SPDLOG_INFO("kill sent to server");
     } else if (vm.count("dir")) {
